+++ cores/epoxy/Arduino.cpp
@@ -21,6 +21,8 @@
 // Arduino methods emulated in Unix
 // -----------------------------------------------------------------------
 
+static uint32_t digitalPinValues = 0;
+
 void yield() {
   usleep(1000); // prevents program from consuming 100% CPU
 }
@@ -29,7 +31,21 @@ void pinMode(uint8_t /*pin*/, uint8_t /*mode*/) {}
 
 void digitalWrite(uint8_t /*pin*/, uint8_t /*val*/) {}
 
-int digitalRead(uint8_t /*pin*/) { return 0; }
+int digitalRead(uint8_t pin) {
+  if (pin >= 32) return 0;
+
+  return (digitalPinValues & (((uint32_t)0x1) << pin)) != 0;
+}
+
+void digitalReadValue(uint8_t pin, uint8_t val) {
+  if (pin >= 32) return;
+
+  if (val == 0) {
+    digitalPinValues &= ~(((uint32_t)0x1) << pin);
+  } else {
+    digitalPinValues |= ((uint32_t)0x1) << pin;
+  }
+}
 
 int analogRead(uint8_t /*pin*/) { return 0; }