+++ CHANGELOG.md
@@ -8,6 +8,9 @@
         * The `pin` parameter must satisfy `0 <= pin < 32`, otherwise
           `digitalReadValue()` is a no-op.
         * See [PR#61](https://github.com/bxparks/EpoxyDuino/pull/61).
+    * Add an empty `EpoxyDuino.h` at the top level to stop warning messages from
+      the Arduino IDE. Fixes
+      [Issue#62](https://github.com/bxparks/EpoxyDuino/issues/62).
 * 1.2.3 (2022-02-24)
     * Rename `unixhostduino_main()` to `epoxyduino_main()`, and make it
       static. No need to expose it publicly.